refactor: resolveMember into util function
