+++ src/commands/moderation/give.ts
@@ -1,7 +1,7 @@
 import Command from '../../lib/structures/Command';
 import { TypicalGuildMessage } from '../../lib/types/typicalbot';
 import { Modes, PermissionsLevels } from '../../lib/utils/constants';
-import { permissionError } from '../../lib/utils/util';
+import { permissionError, resolveMember } from '../../lib/utils/util';
 
 const regex = /(?:(?:(?:<@!?)?(\d{17,20})>?)|(?:(.+)#(\d{4})))?(?:\s+)?(?:(?:<@&)?(\d{17,20})>?|(.+))/i;
 
@@ -25,7 +25,7 @@ export default class extends Command {
 
         const [id, username, discriminator, roleID, roleName] = args;
         const permissions = await message.member.fetchPermissions(true);
-        const member = await this.client.helpers.resolveMember.execute(message, id, username, discriminator, false);
+        const member = await resolveMember(this.client, message, id, username, discriminator, false);
 
         const role = roleID
             ? message.guild.roles.cache.get(roleID)
