+++ src/lib/utils/util.ts
@@ -98,3 +98,30 @@ export const permissionError = (client: TypicalClient, message: TypicalGuildMess
         userTitle: userLevel.title
     });
 };
+
+export const resolveMember = async (client: TypicalClient, message: TypicalGuildMessage, id?: string, username?: string, discriminator?: string, returnSelf = true) => {
+    if (id) {
+        const user = await client.users.fetch(id).catch(console.error);
+        if (!user) return returnSelf ? message.member : null;
+
+        const member = await message.guild.members
+            .fetch(user)
+            .catch(console.error);
+        if (!member) return returnSelf ? message.member : null;
+
+        return member;
+    }
+
+    if (username && discriminator) {
+        await message.guild.members
+            .fetch({ query: username })
+            .catch(console.error);
+
+        const member = message.guild.members.cache.find((m) => m.user.tag === `${username}#${discriminator}`);
+        if (!member) return returnSelf ? message.member : null;
+
+        return member;
+    }
+
+    return returnSelf ? message.member : null;
+};
